rename directory from cur_frames to cru_frames
